Move to hugo.yml
In Hugo 0.110.0 we changed the default configuration base file name to hugo, e.g. hugo.toml. We will still look for config.toml etc., but we recommend you eventually rename it (but you need to wait if you want to support older Hugo versions). The main reason we’re doing this is to make it easier for code editors and build tools to identify this as a Hugo configuration file and project.